Adding HTML and JS files to assets directory
Adding Toshi’s web pages to the app’s asset directory in an attempt to
load the files natively rather than from the web. I’ve guarded the
loading code with a flag - LOAD_HTML_NATIVELY - I turned it on to TRUE
so it should run it locally. Could someone test that has a midi setup?